requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyodbc 
authlib 
itsdangerous
python-jose[cryptography]
httpx
pymongo[srv]
motor
langchain-anthropic